Migrate config-channels
